dotfiles.sh: Readability improvements.
Actually I went to edit this script to fix a bug, only to find out that it wasn't a bug at all...

ln was warning me about not wanting to overwrite the directory ~/.gnupg on my notebook.
I thought I was supposed to change this behaviour, but actually I don't want directories overwritten without being asked for it.