Uninstalling AllChars because of problems with IE.
Apparently AllChars 4 does not work with Internet Explorer 7 under
Windows Vista. No idea why, but IE crashes as soon as I'm trying to
start it, and it works fine when AllChars is not running.

Also, when starting it without AllChars being loaded, it crashes as soon
as it is started.