Mutt: Give up trying to flag mails as read.
If you delete a mail from Google Mail’s IMAP Inbox, it will not be
deleted, but instead “archived” (i.e. still available in the “All Mail”
folder). This is nice, but bears a problem: When you read the mail in
Mutt and delete it afterwards, Mutt will not send the “is read” flag to
the server — because the mail will be deleted anyway (at least that’s
what Mutt thinks).

The only solution to this would be to run <sync-mailbox> (usually bound
to “$”) after reading, but _before_ deleting the message. This is slow
and sucks. Therefore I’m simply ignoring the read/unread marker on my
mails from now on. Since I’m using the Inbox Zero approach, I don’t
really need it anyway: Mails that are in the inbox are not yet handled,
and mails outside of the inbox have already been dealt with.