Use machine branches, not the Makefile.
Before, I would use the res/Makefile and a non-public ".scyprv"
repository to store non-public things (like SSH host names etc). Now, I
am still using a private repository, but it is basically the same as the
public dotscy repository, with the exception that it contains several
additional branches with private data.