Make xterms set TERM=xterm-256color by default
It's understandable that xterm itself defaults to TERM=xterm, even if
256 color mode is compiled in, because the terminfo file(s) on the
system(s) accessed might not know TERM=xterm-256color.

However, it sucks that there is no real possibility to set it to
-256color automatically if it's supported. Because of that, I now
_force_ it to be -256color, even though the system I run this config on
might have an xterm without 256 color mode.

Sigh. I'll fix that as soon as it becomes a problem.