Advanced keyboard config including Apple keyboard
This uses Caps Lock as Compose key and changes the modifier keys right
to the space bar on my Apple keyboard. Additionally, F13 is Print and
F15 is Insert.

Furthermore, the "zap server" option moved out of .xinitrc and into
x-peripherals.sh, which will be called from there anyway. This is
because setxkbmap resets xmodmap and I want to keep confusion and
breakages because of run order problems to a minimum. ;)

Now the next thing I want is to have Apple's special keys on Alt Gr or
something like that…