Add git-find-unignored
You're running "git status" and Git shows you a directory where you are
.gitignore'ing _some_ files, but not all of them. Apparently there's at
least one file somewhere down the tree which is _not_ ignored - but
where? Because Git only shows you the name of the directory where a not
ignored tree starts.

git-find-unignored uses git-check-ignore in combination with find and a
bit of filtering to output you those non-directory items in a certain
directory which are _not_ ignored. Supply it with a directory name as a
parameter and it'll do its magic.

Why only non-directory items, you ask? Because Git ignores empty
directories, so we really don't care about these.