bin/IDE: remove "current"
I wrote this script while having _both_ PhpStorm and IDEA installed, and
"preferred" would link to the preferred IDE, while the "current" link in
the PhpStorm/ or IDEA/ directory would link to its current version.

That's actually nonsense, so I now only have a "preferred" link which
links to the preferred IDE in its desired version.