Add script to fix all that symlink chaos in Windows
Running certain Git commands in my home dir on Windows, e.g. git reset,
leads to Git deleting the contents of directories behind symlinks in
order to make the symlinks “real” symlinks again.

For example, .vim/bundle/solarized is a link to
res/solarized/vim-colors-solarized/autoload. `git windows-symlinks` will
create a Windows directory junction instead of the symlink. Git doesn’t
know about this and will therefore remove the “directory” (rather:
junction) and replace it with a plain text file containing the symlink
target. Worse than that: It deletes the _contents_ of the directory
before removing it, causing the directory that’s being linked to (in the
original submodule) to be empty, causing the submodule to appear as
modified. Sigh.

Therefore, reset all the submodules and then fix the symlinks.

Some way to run this automatically when needed would be great …