Remove audit2010/
In 2010, I decided to basically start my configs from scratch and only
allow code that I had reviewed to really be in my dotfiles. Therefore, I
moved all of the numerous plugins etc. I used to audit2010/ with the
intention of pulling them out again one by one after reviewing.

Well, it turns out that I didn't review much in there, and instead it
ended up as a folder full of cruft that I hadn't used for years. Which
is why I'm now removing it. It will all be still available in the
history and in the "before-2017" tag I just created.